intel/fs: Add a new coverage_mask system value
This represents the actual coverage for the entire pixel, regardless of
whether or not we're shading per-sample.  In theory, we could lower
load_sample_mask_in to this in NIR and avoid the back-end code but
that's a project for another time.

v2: rebase and update on mesa main
v3: rename coverage_mask_intel to coverage_mask to make it generic for
all drivers
Signed-off-by: Rohan Garg <rohan.garg@intel.com>