vbo: extract two methods from compile_vertex_list
They have long list of parameter but it clarifies a bit the purpose of
this code.

While at it, rename buffer_offset: vao_buffer_offset makes it clear
that it's the offset used by the VAO, not the offset where we're going
to write in the buffer.
They can be identical, or vao_buffer_offset can be smaller,
when start_offset is used (see vbo_reuse_same_bo).

Acked-by: Marek Olšák <marek.olsak@amd.com>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/28688>