panfrost: Don't overflow attribute buffer descriptor array
`emit_image_bufs` is given an array of size `N * 2`, where `N` is the
number of attributes used by the shader
(`shader->info.attribute_count`). (It's actually `N * 2 + 1` on Bifrost
but that's not relevant here).

But `emit_image_bufs` writes into that array `M * 2` times, where `M` is
the number of bits set in the shader's image mask
(`panfrost_context.image_mask[<shader type>]`)

If there are images still set in the mask but not used by this shader,
this results in `emit_image_bufs` overflowing its array.

Fix this by increasing the size of the array to allow descriptors for
the entire (possibly sparse, in which case the descriptors are empty)
image mask.

Signed-off-by: Ray Smith <rsmith@brightsign.biz>