nir: add gathering passes that gather which inputs affect specific outputs
The first pass computes which shader input components and which types of
memory loads are used to compute shader outputs. It can be used to query
how data flows within shaders towards outputs.

The second pass uses the first pass to gather which input components are
used to compute pos and clip dist outputs, which input components are used
to compute all other outputs, and which input components are used to
compute both. This will be used by compaction in nir_opt_varyings for
drivers that split TES into a separate position cull shader and varying
shader to make it less likely that the same vec4 inputs are needed in both.