WIP: nir: SWAR vectorizer
I don't know if this is a good idea, mostly posting this to follow up on a
discussion I had with Daniel Schuermann.

This uses the existing vectorization infrastructure to implement i16vec2 packed
math with i32 scalar operations instead, for bitwise ops where we can do this.
This might be more interesting if we could support adds, shifts, etc but that
works only if we can prove there aren't carries/overflows which is challenging.
This also is probably worse than swizzles. But maybe there's something to this?
Or maybe this is all silly? IDK. Posting the MR so I can close it tomorrow and
forget about this little excursion ;)

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>