tu: Make sure tiles being merged are adjacent
Even though we always try to merge a horizontally or vertically adjacent
tile, when we try to merge a vertically adjacent tile it may not
actually be adjacent because it was merged horizontally and the current
tile wasn't or vice versa. We have to detect this and reject merging it.

Fixes: 3fdaad0948c ("tu: Implement bin merging for fragment density map")